+++ clients/web3-wallet/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## 6.0.0 - 2026-06-19
+
+### Added (2)
+
+- `getRfqOrderStatus()` (`GET /api/v1/dex/aggregator/order/{orderId}`)
+- `submitRfqOrder()` (`POST /api/v1/dex/aggregator/order/submit`)
+
+### Changed (4)
+
+- Added parameter `userWalletAddress`
+  - affected methods:
+    - `getAggregatedQuote()` (`GET /api/v1/dex/aggregator/quote`)
+- Added parameter `vendor`
+  - affected methods:
+    - `getErc20ApproveTransaction()` (`GET /api/v1/dex/aggregator/approve-transaction`)
+- Modified response for `getAggregatedQuote()` (`GET /api/v1/dex/aggregator/quote`):
+  - `data`.items: property `approveTarget` added
+  - `data`.items: property `executionMode` added
+  - `data`.items: property `isBest` added
+  - `data`.items: item property `approveTarget` added
+  - `data`.items: item property `executionMode` added
+  - `data`.items: item property `isBest` added
+
+- Modified response for `buildSwapTransaction()` (`GET /api/v1/dex/aggregator/swap`):
+  - `data`: property `executionMode` added
+  - `data`: property `rfq` added
+
 ## 5.0.0 - 2026-06-17
 
 ### Changed (1)